+++ sso/cache.go
@@ -99,31 +99,6 @@ func (c *Cache) GetSSO() *SSOCache {
 	return c.SSO[c.ssoName]
 }
 
-// Expired returns if our Roles cache data is too old.
-// If configFile is a valid file, we check the lastModificationTime of that file
-// vs. the ConfigCreatedAt to determine if the cache needs to be updated
-func (c *Cache) Expired(s *SSOConfig) error {
-	if c.Version < CACHE_VERSION {
-		return fmt.Errorf("Local cache is out of date; current cache version %d is less than %d", c.Version, CACHE_VERSION)
-	}
-
-	// negative values disable refresh
-	if s.settings.CacheRefresh <= 0 {
-		return nil
-	}
-
-	ttl := s.settings.CacheRefresh * 60 * 60 // convert hours to seconds
-	cache := c.GetSSO()
-	if cache.LastUpdate+ttl < time.Now().Unix() {
-		return fmt.Errorf("Local cache is out of date; TTL has been exceeded.")
-	}
-
-	if s.CreatedAt() > c.ConfigCreatedAt {
-		return fmt.Errorf("Local cache is out of date; config.yaml modified.")
-	}
-	return nil
-}
-
 func (c *Cache) CacheFile() string {
 	return c.settings.cacheFile
 }
@@ -271,7 +246,7 @@ func (c *Cache) deleteOldHistory() {
 
 // Refresh updates our cached Roles based on AWS SSO & our Config
 // but does not save this data!
-func (c *Cache) Refresh(sso *AWSSSO, config *SSOConfig, ssoName string) error {
+func (c *Cache) Refresh(sso *AWSSSO, config *SSOConfig, ssoName string, threads int) error {
 	// Only refresh once per execution
 	if c.refreshed {
 		return nil
@@ -306,7 +281,7 @@ func (c *Cache) Refresh(sso *AWSSSO, config *SSOConfig, ssoName string) error {
 	c.SSO[ssoName].Roles = &Roles{}
 
 	// load our AWSSSO & Config
-	r, err := c.NewRoles(sso, config)
+	r, err := c.NewRoles(sso, config, threads)
 	if err != nil {
 		return err
 	}
@@ -418,7 +393,7 @@ func (c *Cache) GetRole(arn string) (*AWSRoleFlat, error) {
 
 // Merges the AWS SSO and our Config file to create our Roles struct
 // which is defined in cache_roles.go
-func (c *Cache) NewRoles(as *AWSSSO, config *SSOConfig) (*Roles, error) {
+func (c *Cache) NewRoles(as *AWSSSO, config *SSOConfig, threads int) (*Roles, error) {
 	r := Roles{
 		SSORegion:     config.SSORegion,
 		StartUrl:      config.StartUrl,
@@ -427,7 +402,7 @@ func (c *Cache) NewRoles(as *AWSSSO, config *SSOConfig) (*Roles, error) {
 		ssoName:       config.settings.DefaultSSO,
 	}
 
-	if err := c.addSSORoles(&r, as); err != nil {
+	if err := c.addSSORoles(&r, as, threads); err != nil {
 		return &Roles{}, err
 	}
 
@@ -499,7 +474,7 @@ func processSSORoles(roles []RoleInfo, cache *SSOCache, r *Roles) {
 }
 
 // addSSORoles retrieves all the SSO Roles from AWS SSO and places them in r
-func (c *Cache) addSSORoles(r *Roles, as *AWSSSO) error {
+func (c *Cache) addSSORoles(r *Roles, as *AWSSSO, threads int) error {
 	cache := c.GetSSO()
 
 	accounts, err := as.GetAccounts()
@@ -519,9 +494,9 @@ func (c *Cache) addSSORoles(r *Roles, as *AWSSSO) error {
 	// Per #448, doing this serially is too slow for many accounts.  Hence,
 	// we'll use a worker pool.
 	if len(accounts) > 0 {
-		workers := 1
-		if c.settings.Threads > 0 {
-			workers = c.settings.Threads
+		workers := c.settings.Threads
+		if threads > 0 {
+			workers = threads
 		}
 		if workers > len(accounts) {
 			workers = len(accounts)